attached_assets/extracted/smart/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.39
Werkzeug==2.2.2
google-generativeai==0.7.2
python-dotenv==1.0.0